code imports in more dynamic way, separate from entry point
