+++ recipe/bld.bat
@@ -4,7 +4,10 @@ COPY conda_src\conda\utils.py %SP_DIR%\conda\utils.py
 COPY menuinst_src\menuinst\__init__.py %SP_DIR%\menuinst\__init__.py
 COPY menuinst_src\menuinst\win32.py %SP_DIR%\menuinst\win32.py
 
-:: -F is to create a single file
-pyinstaller -F -n conda.exe conda.exe.py
+:: This is ordinarily installed by the installer itself, but since we are building for a
+:: standalone and have only an env, not an installation, include it here.
+COPY constructor\constructor\nsis\_nsis.py %PREFIX%\Lib\_nsis.py
+
+pyinstaller conda.exe.spec
 MKDIR %PREFIX%\standalone_conda
 MOVE dist\conda.exe %PREFIX%\standalone_conda\conda.exe
